+++ sqlservice/__init__.py
@@ -12,5 +12,5 @@ from .__pkg__ import (
 
 from .client import SQLClient
 from .core import destroy, make_identity, save, transaction
-from .model import ModelBase, declarative_base
+from .model import ModelBase, as_declarative, declarative_base
 from .query import SQLQuery